telegram-stock-bot_alpha/requirement.txt
python-telegram-bot==20.7
requests
pandas
matplotlib
python-dotenv
finnhub-python==2.4.15 